move runfileconfig
